Rename types to constants
